chore: move some tests to engine crate
