+++ crates/rpc/rpc-engine-api/Cargo.toml
@@ -30,6 +30,7 @@ jsonrpsee-core = "0.18"
 tracing.workspace = true
 
 [dev-dependencies]
+reth-rlp.workspace = true
 reth-interfaces = { workspace = true, features = ["test-utils"] }
 reth-provider = { workspace = true, features = ["test-utils"] }
 reth-payload-builder = { workspace = true, features = ["test-utils"] }
