+++ crates/rpc/rpc-engine-api/tests/it/main.rs
@@ -1,0 +1,3 @@
+mod payload;
+
+fn main() {}
